pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Aas Paas MCP Server for the Puch AI Hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.2",
    "markdownify>=1.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
    # --- Our required libraries are now officially part of the project ---
    "google-generativeai",
    "requests",
    "uvicorn",
    "pydantic-settings",
    "beautifulsoup4",
]